Add analytics

+++ app/controllers/searches_controller.rb
@@ -2,6 +2,15 @@ class SearchesController < ApplicationController
   def index
     # display all saved searches for the current user using ip address tracking
     @searches = all_user_searches
+    @counts = Search.group(:text).count
+  end
+
+  def analytics
+    @counts = Search.group(:text).where(user: request.remote_ip).count
+    respond_to do |format|
+      format.json { render json: Search.group(:text).count }
+      format.html { render partial: 'searches/analytics', status: :ok, locals: { counts: @counts }}
+    end
   end
 
   # route for saving searches it saves the search term and user (via user ip address)
@@ -16,14 +25,9 @@ class SearchesController < ApplicationController
 
     respond_to do |format|
       if ignore_char_deletion?(last_search, @search)
-        format.turbo_stream
-        format.json { render json: @searches, status: :ok }
-        # add an html format that displays all the searches and displays a partial in searces/search
-        format.html { render partial:"searches/save", status: :ok}
+        format.html { render partial: 'searches/save', status: :ok }
       elsif @search.save
-        format.turbo_stream
-        format.json { render json: @searches, status: :created }
-        format.html { render partial: "searches/save", status: :ok }
+        format.html { render partial: 'searches/save', status: :ok }
 
       else
         format.html { render :new, status: :unprocessable_entity }
@@ -35,7 +39,7 @@ class SearchesController < ApplicationController
   private
 
   def ignore_char_deletion?(last_search, search)
-    last_search && last_search.text.strip.start_with?(search.text.strip)
+    last_search && last_search.text.strip.start_with?(search.text.strip) && last_search.text != search.text
   end
 
   def all_user_searches
@@ -45,7 +49,7 @@ class SearchesController < ApplicationController
   def remove_last_saved_search(last_search, search)
     id = nil
 
-    if last_search && search.text.strip.start_with?(last_search.text.strip)
+    if last_search && search.text.strip.start_with?(last_search.text.strip) && last_search.text != search.text
       id = last_search.id
       last_search.destroy
     end
